Have tourItemsReducer be a separate file
